"Get resource type" is triggering RDF error
We call get resource type with the descMetadata content stream. This in turn
parses the stream using rdf/ntriples, which in some case, was throwing an
error, stating that "private method readline' called for nil:NilClass". I added
a test that the key exists prior to trying to access and parse each datastream.