Adjust how the agent is parsed
The code was stashing the _complete_ log line in the agent field and
then, later on, parsing out the agent. Now it parses the agent at the
very beginning.

This changes the test since there were a few lines including referrals
from yahoo that were being (wrongy) dropped using the prior method. The
new method is more precise.