Removing brakeman identified vulnerabilities
Rails 4.2.6 had vulnerabilities that are addressed with a Rails
version update. In opening the Rails version, I then encountered
an upstream `rake` issue and froze to a compatable version.

This application is an odd duck in that we don't need the
full Rails ecosystem for report generation, however it is
what we have.