Updating to get local tests running
There were a few components. First, ensuring that we are using an
updated Ruby version (as mirroed by CurateND). Second, locking the
mysql2 gem to that of Curate and to a version of mysql2 that is
alignment with the Rails version. Third, updating the README to
include guidance on running tests. And Fourth, updating the database.yml
file to include localhost so `bundle exec rake` should work out of the
box.